backend/requirements.txt
# Web framework
fastapi==0.115.12
uvicorn[standard]==0.34.2

# ORM & Database
sqlmodel==0.0.24
asyncpg==0.30.0
alembic==1.15.2
greenlet==3.2.1

# Authentication & Security
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.4.0
python-multipart==0.0.20

# Rate limiting
slowapi==0.1.9

# Validation
pydantic[email-validator]==2.11.2
pydantic-settings==2.14.0

# Environment
python-dotenv==1.1.0

# HTTP client (MercadoPago, webhooks)
httpx==0.28.1
mercadopago==2.2.2
